removing test-support module in favor of docker-compose
