+++ modules/core/src/main/scala/dev/profunktor/redis4cats/connection/RedisClient.scala
@@ -36,13 +36,13 @@ object RedisClient {
 
     val release: RedisClient => F[Unit] = client =>
       F.info(s"Releasing Redis connection: $uri") *>
-          JRFuture.fromCompletableFuture(F.delay(client.underlying.shutdownAsync())).void
+        JRFuture.fromCompletableFuture(F.delay(client.underlying.shutdownAsync())).void
 
     (acquire, release)
   }
 
   private[redis4cats] def acquireAndReleaseWithoutUri[F[_]: Concurrent: ContextShift: Log]
-      : F[(F[RedisClient], RedisClient => F[Unit])] =
+    : F[(F[RedisClient], RedisClient => F[Unit])] =
     F.delay(RedisURI.fromUnderlying(new JRedisURI())).map(acquireAndRelease(_))
 
   def apply[F[_]: Concurrent: ContextShift: Log](uri: => RedisURI): Resource[F, RedisClient] = {
